Refactor: Move engine-specific classes to source package (PR 1)
Move MySql, Postgres, and Oracle ChangeEventContext and ChangeEventSequence
classes, along with their unit tests, to their respective sub-packages
under v2/datastream-to-spanner/src/main/java/.../datastream/source/{engine}/.

This is the first PR in a two-part series to decouple the datastream-to-spanner
template from specific source engines. This PR focuses purely on moving files
and resolving compilation/visibility issues to ensure Git preserves history
as 100% renames.

- Moved 6 source files and 6 test files to engine-specific packages.
- Made moved classes public to allow factory access.
- Extracted ChangeEventSequenceComparisonException to its own public file.
- Made ChangeEventContext.generateShadowTableMutation protected to allow cross-package overriding.
- Made ChangeEventConvertor.changeEventToShadowTableMutationBuilder public.
- Made ChangeEventSequence constructor protected.
- Made getters in sequence classes public for test access.
- Added package-info.java for new packages.
- Fixed import orders and added missing imports.

TAG=agy
CONV=b7fec87d-3652-43db-a0d1-64773a054415
